pyproject.toml
[tool.poetry]
name = "atlas"
version = "0.1.0"
description = "Atlas — personal ops agent (backend)"
authors = ["Hamza"]
readme = "README.md"
packages = [{ include = "atlas", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.115"
uvicorn = { extras = ["standard"], version = ">=0.32" }
pydantic = ">=2"
pydantic-settings = ">=2"
sqlalchemy = { extras = ["asyncio"], version = ">=2.0" }
asyncpg = ">=0.30"
alembic = ">=1.14"
structlog = ">=24"
python-dotenv = ">=1"
google-generativeai = ">=0.8"

[tool.poetry.group.dev.dependencies]
pytest = ">=8"
pytest-asyncio = ">=0.24"
httpx = ">=0.28"
ruff = ">=0.8"
mypy = ">=1.13"
pre-commit = ">=4"
factory-boy = ">=3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "SIM", "RUF", "T20"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"migrations/**" = ["INP001"]
"scripts/**" = ["T201"]
"tests/**" = ["S101"]
"tests/conftest.py" = ["E402"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_any_unimported = false
ignore_missing_imports = false
mypy_path = "src"
packages = ["atlas"]

[[tool.mypy.overrides]]
module = [
    "alembic.*",
    "factory_boy.*",
    "structlog.*",
    "uvicorn.*",
    "dotenv.*",
    "google.*",
    "google.generativeai.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
